+++ client/src/main/java/com/guilherme/todo_app/controller/TokenController.java
@@ -1,5 +1,6 @@
-package com.guilherme.todo_app.security.jwt;
+package com.guilherme.todo_app.controller;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
@@ -8,9 +9,13 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
-import com.guilherme.todo_app.model.users.User;
-import com.guilherme.todo_app.model.users.UserRepository;
+import com.guilherme.todo_app.model.jwt.JWTTokenRequest;
+import com.guilherme.todo_app.model.jwt.JWTTokenResponse;
+import com.guilherme.todo_app.model.jwt.TokenService;
+import com.guilherme.todo_app.model.user.User;
+import com.guilherme.todo_app.model.user.UserRepository;
 
 @RestController
 @RequestMapping("/auth")
@@ -19,6 +24,7 @@ public class TokenController {
     private final TokenService TokenService;
     private final UserRepository userRepository;
     private final AuthenticationManager authenticationManager;
+
     private final BCryptPasswordEncoder encoder = new BCryptPasswordEncoder(12);
     
     public TokenController(TokenService TokenService, AuthenticationManager authenticationManager, UserRepository userRepository) {
@@ -30,26 +36,17 @@ public class TokenController {
     @PostMapping("/login")
     public ResponseEntity<?> generateToken(@RequestBody JWTTokenRequest jwtTokenRequest) {
 
-        // Optional<User> user = userRepository.findByEmail(jwtTokenRequest.getEmail());
-        // if(user.isEmpty()) {
-        //     return ResponseEntity.of(user);
-        // }
-
-        System.out.println("1");
         var authenticationToken =
             new UsernamePasswordAuthenticationToken(
                 jwtTokenRequest.getUsername(), 
                 jwtTokenRequest.getPassword());
 
-        System.out.println("2");
         var authentication =
             authenticationManager.authenticate(authenticationToken);
 
-        System.out.println("3");
         var token =
             TokenService.generateToken(authentication);
 
-        System.out.println("4");
         return ResponseEntity.ok(new JWTTokenResponse(token));
 
     }
@@ -57,11 +54,16 @@ public class TokenController {
     @PostMapping("/register")
     public ResponseEntity<?> registerUser(@RequestBody User user) {
 
+        if(userRepository.findByUsername(user.getUsername()).isPresent()) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+            //return new ResponseEntity<>("E-Mail já cadastrado.", HttpStatus.CONFLICT);
+        }
+        
         user.setPassword(encoder.encode(user.getPassword()));
         userRepository.save(user);
         System.out.println(user);
 
-        return null;
+        return new ResponseEntity<>(HttpStatus.OK);
     }
     
 }
