+++ client/src/main/java/com/guilherme/todo_app/model/user/User.java
@@ -1,4 +1,4 @@
-package com.guilherme.todo_app.model.users;
+package com.guilherme.todo_app.model.user;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
